+++ packages/components/src/VirtualizedList/CellDatetime/CellDatetime.component.js
@@ -3,11 +3,12 @@ import React from 'react';
 import classnames from 'classnames';
 import isEqual from 'lodash/isEqual';
 import pick from 'lodash/pick';
-import { distanceInWordsToNow, format } from 'date-fns';
+import distanceInWordsToNow from 'date-fns/distance_in_words_to_now';
+import format from 'date-fns/format';
 import isValid from 'date-fns/is_valid';
 import parse from 'date-fns/parse';
 import { withTranslation } from 'react-i18next';
-import talendUtils from '@talend/utils';
+import { date as dateUtils } from '@talend/utils';
 
 import I18N_DOMAIN_COMPONENTS from '../../constants';
 import getDefaultT from '../../translate';
@@ -28,7 +29,7 @@ export function computeValue(cellData, columnData, t) {
 			});
 		} else if (columnData.mode === 'format') {
 			if (columnData.timeZone) {
-				return talendUtils.date.formatToTimeZone(cellData, columnData.pattern || DATE_TIME_FORMAT, {
+				return dateUtils.formatToTimeZone(cellData, columnData.pattern || DATE_TIME_FORMAT, {
 					timeZone: columnData.timeZone,
 				});
 			}
@@ -40,15 +41,19 @@ export function computeValue(cellData, columnData, t) {
 }
 
 export function getTooltipLabel(cellData, columnData) {
-	if(typeof columnData.getTooltipLabel === 'function') {
+	if (typeof columnData.getTooltipLabel === 'function') {
 		return columnData.getTooltipLabel(cellData);
 	}
 	if (columnData.mode === 'ago') {
 		let tooltipLabel = '';
 		if (columnData.timeZone) {
-			tooltipLabel = talendUtils.date.formatToTimeZone(cellData, columnData.pattern || DATE_TIME_FORMAT, {
-				timeZone: columnData.timeZone,
-			});
+			tooltipLabel = dateUtils.formatToTimeZone(
+				cellData,
+				columnData.pattern || DATE_TIME_FORMAT,
+				{
+					timeZone: columnData.timeZone,
+				},
+			);
 		} else {
 			tooltipLabel = format(cellData, columnData.pattern || DATE_TIME_FORMAT);
 		}